wip Battle create endpoint and extract PokemonSelector
