+++ assets/js/utils/api.js
@@ -1,5 +1,9 @@
 import axios from 'axios';
-import Urls from './urls';
+import { getCookieValue, Urls } from '.';
+
+const apiWrapper = axios.create({
+  headers: { 'X-CSRFToken': getCookieValue('csrftoken') },
+});
 
 export default class Api {
   static getBattleDetails(id) {
@@ -33,6 +37,13 @@ export default class Api {
   static loadPokemonList() {
     const url = Urls['api-pokemon:list']();
     return axios.get(url)
+      .then(response => response.data)
+      .catch(error => new Error(error));
+  }
+
+  static createBattle(values) {
+    const url = Urls['api-battles:create-battle']();
+    return apiWrapper.post(url, values)
       .then(response => response.data)
       .catch(error => new Error(error));
   }
