+++ battles/endpoints.py
@@ -1,10 +1,12 @@
 from django.db.models import Q
 
-from rest_framework import generics
+from rest_framework import generics, permissions
+from rest_framework.response import Response
 
 from battles.models import Battle
 from battles.permissions import IsInBattle
 from battles.serializers import BattleListSerializer, BattleSerializer
+from pokemons.serializers import PokemonSerializer
 
 
 class BattleDetailsEndpoint(generics.RetrieveAPIView):
@@ -22,3 +24,24 @@ class BattleListEndpoint(generics.ListAPIView):
         return Battle.objects.filter(
             Q(creator=user) | Q(opponent=user)
         )
+
+
+class BattleCreateEndpoint(generics.CreateAPIView):
+    serializer_class = BattleSerializer
+    permission_classes = (permissions.IsAuthenticated, )
+
+    def post(self, request, *args, **kwargs):
+        data = request.data
+        battle = dict(
+            opponent=data.opponent,
+            creator=request.user
+        )
+        battle = BattleSerializer(data=battle)
+        battle.is_valid()
+        battle.save()
+        creator_team = dict(
+            battle_related=battle.id,
+            pokemons=PokemonSerializer(data=data.team, many=True)
+        )
+        # WIP
+        return Response(creator_team)
